Update PULL
For the Idle for Three Second termination portion (Task 2), there is an issue somewhere along the line. The gyro may not be properly calibrated through the code due to how it sits on the bot (not directly perpendicular to the force of gravity, but close). The task may also not terminate the entire program properly, which is another coding issue. Other than that, the program runs how it should, and the bot maintains a straightforward path down the pull field.